+++ src/main/java/com/USWRandomChat/backend/chat/api/ChatController.java
@@ -2,8 +2,7 @@ package com.USWRandomChat.backend.chat.api;
 
 import com.USWRandomChat.backend.chat.chatDTO.ChatMessage;
 import com.USWRandomChat.backend.chat.domain.ChatRoom;
-import com.USWRandomChat.backend.chat.repository.ChatRepository;
-import com.USWRandomChat.backend.chat.service.RedisPublisher;
+import com.USWRandomChat.backend.chat.service.ChatRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.data.redis.listener.ChannelTopic;
@@ -25,7 +24,6 @@ public class ChatController {
         //메시지 발송 시 /pub/chat/message
         //메시지 수신 시 /sub/chat/room/방 ID
         if (ChatMessage.MessageType.JOIN.equals(message.getType())) {
-            message.setSender("[알림]");
             message.setMessage(message.getSender() + "님이 입장하였습니다.");
         }
         //webSocket에 발행된 메시지를 redis로 발행한다.(publish)
